third_party/blink/web_tests/wpt_internal/webgl/draw-element-image/drawing-display-none-fails.html

<!DOCTYPE html>
<title>Drawing an element with display:none should fail</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<canvas id="canvas_3d" layoutsubtree width="100" height="100" style="background: green;">
  <div id="canvas_3d_div" style="width: 100px; height: 100px; background: red;"></div>
</canvas>

<script>
  promise_test(async function() {
    canvas_3d_div.getBoundingClientRect();
    canvas_3d_div.style.display = 'none';
    await new Promise(requestAnimationFrame);
    await new Promise(setTimeout);
    const context_gl = canvas_3d.getContext('webgl2');
    const tex = context_gl.createTexture();
    context_gl.bindTexture(context_gl.TEXTURE_2D, tex);
    const internalformat = context_gl.RGBA8;
    assert_throws_dom(
      "InvalidStateError",
        () => context_gl.texElementImage2D(
            context_gl.TEXTURE_2D, internalformat, canvas_3d_div),
      'texElementImage2D cannot draw display:none elements.');
  }, 'texElementImage2D with display:none element');
</script>
